refactor: 사다리 출력 문자열 생성 로직 클래스 분리
